added support for coloring mRNA, blastz mRNA, and xenoMrna by cds.
Also made some improvements to the regular cds coloring of genePred
tracks. Most notably is the color change, it now uses the track color
and a lighter shade and only colors if BOTH ends of the cds are
complete.

If available, the extra frame information is used (but you still have
to select it in the options menu). The only applies to the NCBI genes
track currently.