added option for codon coloring to let users turn it on for 3 psl tracks. This doesn't make it on by default. For that you need to add
cdsDrawDefault genomic\ codons

for example to turn "genomic codons" mode on. An example for the mRNA
track would be

cdsDrawOptions enabled
cdsDrawDefault mRNA\ codons

The options can be disabled (so the user can't change them) but still
be set to anything you want in the trackDb. Only "genomic codons" or
"off" are valid modes for genePred tracks. genePred options are enabled
by default and all others tracks must be explicitly enabled.