Moved filterOnRelease into tdbsForDbPath, calling it just after we load stanzas from .ra file(s) at each level, instead of calling it after tdbsForDbPath.  The reason for filtering earlier is that stanzas for incompatible releases are hard not to trip on somewhere in the complex logic of merging records.
The specific case that was failing, mentioned in MLQ #1963:
snp131 has no release specified in the top-level trackDb.ra,
but in human/hg19/trackDb.ra, it has two overrides, one for release alpha
and the other for beta,public.  The beta,public one was processed before
the alpha one.  Since the top-level had no release, it didn't clash with
beta,public, so the records were merged, leaving a singe beta,public
record.  When we encountered the alpha record, it was incompatible with
the merged top-level+beta,public record so it was not merged and had
only override params, not top-level.  Since it was missing the type
param, it failed the doRecordChecks test.  Filtering out the record
for other releases allowed no-release and alpha to merge properly.