Fixing bug reported by user in MLQ #11775: VCF that is sorted by start position might not be sorted by UCSC-start position, specifically when at the same VCF start position there is an indel followed by a substitution.  VCF start coords for indels are actually one base to the left of the actual inserted or deleted bases.  So we internally adjust those to be one base to the right, which can make them out-of-order when followed by a non-indel variant.
To restore sorting order, indels are now saved aside in a queue until
we know it's safe to let them out because either
a) we got a substitution whose start is greater than the indel start
b) we got an indel whose start is greater than the queued indel start
c) we hit EOF.
When condition (a) or (b) is met, we have to do something with the
new record while the indel queue drains, so it's saved aside in the
nextPosQ (which never holds more than one row).  If it's (b), then
when indelQ is empty, we move the nextPosQ indel to indelQ because
we can't let that indel out until one of the 3 conditions is met.

fixes #11787