It turns out that although it compiles, it does not actually work unless one turns off COLOR32 setting in common.mk, which is not going to happen. I have a local development branch where this works. Alternatives such as converting color values from 32-bit to 256 colors for gif saving could be explored in the future.
Revert "Changing output of phyloGif back go actual gif instead of png.  I first tried changing stuff to be png, but there were many references in html files that would be a real pain to change back. Besides, the gif code is stable and works well. There is no patents on it anymore, and no reason to not use it when it is convenient, which it is here."

This reverts commit 6f8f179499a5fccd49717db4af56d7aec5c9b26c.