In code review, Max pointed out that it's non-standard to conditionally include header files in the middle of a .c file.  While looking at it, I realized that there were a lot of unnecessary #ifdef's for KNETFILE_HOOKS in general, all because there was not yet an #else implementation of knetUdcInstall, which can be a no-op if KNETFILE_HOOKS is not used. So I yanked out a lot of unnecessary old stuff, and tested basic bam and vcf track operations with a few settings:
USE_SAMTABIX   USE_BAM   USE_TABIX   KNETFILE_HOOKS
         1     (these tree are all implied by USE_SAMTABIX)
         0        1          0       0
         0        1          0       1
         0        0          1       1
USE_TABIX without KNETFILE_HOOKS doesn't work because the tabix lib
wants to download .tbi files into the current working directory,
cgi-bin, and we don't allow that (and shouldn't!).  If we were going
to support that, we could add a change-dir solution as in bamFile.c,
but I don't think we need to support it now that there's samtabix.

refs #6235, refs #12850