Built snp142 tracks on hg19 & hg38.
Checked for missing rs_fasta and there wasn't any this time!  dbSNP provided
a file of 1000 Genomes Phase3 indels that changed position and/or allele;
the file had some issues and didn't come with rs_fasta changes so I decided
to leave it as-is.

hg19 still had incorrect weights for variants that uniquely mapped to a
chrom and its alt(s); tweaked those to 1 when we didn't see MultipleAlignments.
hg38 was OK though.

refs #14189