Adding dbDbTaxonomy.pl along with makeDb/doc instructions for generating/updating two new JS files that contain phylogenetic trees for hgwdev's and rr's species.
dbDbTaxonomy.pl uses taxIds extracted from hgcentral(test) to build a phylogenetic tree using NCBI Taxonomy database dump files.
It finds ancestors of the hgcentral species' taxIds to build a tree, and adds node names to ancestor nodes.
It assigns priorities to nodes based on a list of 'favored species' in the order we'd like to see them listed when the tree is traversed.
Then it dumps out Javascript encoding the tree as recursive Arrays.  The JS output can be included as a script in HTML.

refs #15277