Watch out for invalid data and unencoded newlines when reading in file uploaded as a saved session.  refs #22638, #22637
cartLoadSettings did absolutely no checking that its input looked anything like cart settings, resulting in giant useless carts in which custom track files, hgTracks HTML and other arbitrary files apparently were parsed as cart settings.  Now we do quite a bit of checking in cartLoadSettingsFromUserInput, and can warn the user if it looks like they are attempting to upload a custom track as a saved session.

The unencoded newlines are our fault -- since 2010, we have been writing out saved session files without encoding newlines in setting values (pasted in by the user).  Then those newlines can be mistaken for the newlines that we use to separate settings.  Oops!  Use a heuristic that holds up so far to determine whether the newlines should have been encoded: they seem to occur in specific variables (like hgta_pastedIdentifiers), and those variables always appear in the middle of an alphabetically sorted cluster of variables with the same prefix (like hgta_).

And of course, we now write out saved session files with encoded newlines, which we decode when reading back in.

We even have to deal with carts that have mixed types of newlines from various garbage sources, using lineFileCarefulNewlines().