Added -perSeqMax=file option to gfServer, for assemblies with many alt/fix sequences.  refs #23112
gfServer's default maxDnaHits=100 is exceeded easily for transcripts like NM_020535.3 that have clump matches to many spots on chr19 and its dozens of _alts, so only the largest exon matches make the cutoff and we lose the best full transcript alignments.  Passing in -perSeqMax with a list of _alt and _fix filename:seq names enables us to apply the limit separately to each alt (maxDnaHits/2 per alt), so full transcript matches are found.

There can be many more results for hgBlat to work through, which slows it down, but this should apply only when there are many clump matches to many alts.

In the normal case in which a transcript has relatively few hits across the genome, and none to alts, hgBlat's performance is not affected.