dbSnp153: When different freq alleles can be expanded by different amounts, pad out the smaller range & alleles for consistency.  refs #23283 One variant (rs782394990) was dropped because of that quirk, but now no variants are dropped.  :)  A few ucscNotes increased by 1 or 2. Some of the counts of various ucscNotes were a little out of date due to recent work on adding freq notes instead of dropping; up to date now.
I was all proud of this but it's a dead end.
One b153 variant, rs782394990, errored out because its freq alleles couldn't be expanded
by the same amount.  The insertions of pure A's could be expanded to a larger range than
an ins that included a T in addition to A's.  So I siezed upon the problem of independently
trimmed and expanded SPDIs resulting in inconsistent dels, and set about trimming by list
instead of trimming each SPDI separately.  I would restrict the expansion to the minimum
range by which any allele could be shifted.  That would fix the problem of independent
del's for rs782394990.

However, there was a problem.  A variant that includes both an ins and a del that could be
expanded to the same range, e.g. rs201454468 with alleles ref=T, alt="", alt=TT, cannot
be trimmed as a list because it already has "".  T/"" is already minimal.  However, T/TT
is not already minimal and needs to be minimized before expansion.  So no expansion was
performed, and we ended up with inconsistent alleles between different sources because
one was expanded (it had only the del) and the other wasn't.  No good.

I think that's why, although I was trimming listwise already, I did another trim just before
trying to expand.  Is there even a point in trimming listwise?

So... back to individual trimming of each SPDI.  And individual expansion.  But what do we
do about the 1 in 680M case of rs782394990?  I think we should detect inconsistent dels --
and then pad out all dels and inses to the maximal range.  So for the allele that could be
shifted 2 bases less than the others because it included a T, just pad its del and ins with
genomic bases.

OK, now it works.  Squash branch into master...